chore: Move Rego signatures to top level signatures.
