+++ signatures/rego/anti_debugging_ptraceme.rego
