+++ signatures/rego/dropped_executable.rego
