+++ signatures/rego/dropped_executable_test.rego
