+++ signatures/rego/fileless_execution.rego
