+++ signatures/rego/fileless_execution_test.rego
