+++ signatures/rego/illegitimate_shell.rego
