+++ signatures/rego/illegitimate_shell_test.rego
