+++ signatures/rego/kernel_module_loading.rego
