+++ signatures/rego/kernel_module_loading_test.rego
