+++ signatures/rego/kubernetes_certificate_theft_attempt.rego
