+++ signatures/rego/ld_preload.rego
