+++ signatures/rego/ld_preload_test.rego
