+++ signatures/rego/regosig/aio.go
