+++ signatures/rego/regosig/aio.rego
