+++ signatures/rego/regosig/aio_test.go
@@ -6,10 +6,11 @@ import (
 	"os"
 	"testing"
 
+	"github.com/aquasecurity/tracee/signatures/rego/regosig"
+
 	"github.com/aquasecurity/tracee/signatures/signaturestest"
 
 	tracee "github.com/aquasecurity/tracee/pkg/external"
-	"github.com/aquasecurity/tracee/tracee-rules/signatures/rego/regosig"
 	"github.com/aquasecurity/tracee/tracee-rules/types"
 	"github.com/open-policy-agent/opa/compile"
 	"github.com/stretchr/testify/assert"
