+++ signatures/rego/regosig/common_test.go
