+++ signatures/rego/regosig/mapper.go
