+++ signatures/rego/regosig/mapper_test.go
