+++ signatures/rego/regosig/traceerego.go
