+++ tracee-rules/benchmark/signature/rego/signatures.go
@@ -3,9 +3,10 @@ package rego
 import (
 	_ "embed"
 
+	"github.com/aquasecurity/tracee/signatures/rego/regosig"
+
 	"github.com/open-policy-agent/opa/compile"
 
-	"github.com/aquasecurity/tracee/tracee-rules/signatures/rego/regosig"
 	"github.com/aquasecurity/tracee/tracee-rules/types"
 )
